[1/2] Left handed navbar during landscape mode
This patch allows for the navbar to be moved to the left side of the
screen during landscape mode. Left handed people are more used to this
option.

This patch must go with another patch that changes framework/base

PS1 : add check for navbar dependency

Screenshot: http://i.imgur.com/jWqJcEV.png

Change-Id: I04b2be98b9ffce96b4c36c2e3eae398a9e99562f
Signed-off-by: hany <hany.alsamman@gmail.com>