Settings: Show heads up at the bottom of the screen (2/2) Due of user requests and indeed especially on hw key devices for a better workflow we add an option to show heads up notifications on the bottom instead of the top.
Animation we leave as is due that if it slide from the bottom into the screen it happens
to often that the user accidentaly interacts with it. If it comes from the top of the
container we add, the user has enough time to realize it. Sometimes a ms can make a big difference.